refactor(py-sdk): break out different classes into their own modules
updates to some of the experience of using validators
